moved admin-helpers and updated pathes
