+++ web/src/app/admin/bot/[id]/page.tsx
@@ -6,7 +6,7 @@ import { ErrorCallout } from "@/components/ErrorCallout";
 import { DocumentSet, SlackBotConfig } from "@/lib/types";
 import { Text } from "@tremor/react";
 import { BackButton } from "@/components/BackButton";
-import { Persona } from "@/assistants/interfaces";
+import { Persona } from "@/lib/admin/interfaces";
 import { InstantSSRAutoRefresh } from "@/components/SSRAutoRefresh";
 
 async function Page({ params }: { params: { id: string } }) {
