+++ web/src/components/chatPageComponents/lib.tsx
@@ -17,7 +17,7 @@ import {
   StreamingError,
   ToolRunKickoff,
 } from "./interfaces";
-import { Persona } from "@/app/admin/assistants/interfaces";
+import { Persona } from "@/lib/admin/interfaces";
 import { ReadonlyURLSearchParams } from "next/navigation";
 import { SEARCH_PARAM_NAMES } from "./searchParams";
