+++ web/src/components/chatPageComponents/modal/ChatModalAssistantsTab.tsx
@@ -1,4 +1,4 @@
-import { Persona } from "@/app/admin/assistants/interfaces";
+import { Persona } from "@/lib/admin/interfaces";
 import { Bubble } from "@/components/Bubble";
 import { useChatContext } from "@/components/context/ChatContext";
 import { getFinalLLM } from "@/lib/llm/utils";
