+++ web/src/components/context/ChatContext.tsx
@@ -9,7 +9,7 @@ import {
   ValidSources,
 } from "@/lib/types";
 import { ChatSession } from "@/components/chatPageComponents/interfaces";
-import { Persona } from "@/app/admin/assistants/interfaces";
+import { Persona } from "@/lib/admin/interfaces";
 import { LLMProviderDescriptor } from "@/app/admin/models/llm/interfaces";
 import { Folder } from "@/components/chatPageComponents/folders/interfaces";
