+++ web/src/lib/admin/helpers.ts
@@ -1,42 +1,4 @@
-import { Persona, Prompt, StarterMessage } from "./interfaces";
-
-interface PersonaCreationRequest {
-  name: string;
-  description: string;
-  system_prompt: string;
-  task_prompt: string;
-  document_set_ids: number[];
-  num_chunks: number | null;
-  include_citations: boolean;
-  is_public: boolean;
-  llm_relevance_filter: boolean | null;
-  llm_model_provider_override: string | null;
-  llm_model_version_override: string | null;
-  starter_messages: StarterMessage[] | null;
-  users?: string[];
-  groups: number[];
-  tool_ids: number[]; // Added tool_ids to the interface
-}
-
-interface PersonaUpdateRequest {
-  id: number;
-  existingPromptId: number | undefined;
-  name: string;
-  description: string;
-  system_prompt: string;
-  task_prompt: string;
-  document_set_ids: number[];
-  num_chunks: number | null;
-  include_citations: boolean;
-  is_public: boolean;
-  llm_relevance_filter: boolean | null;
-  llm_model_provider_override: string | null;
-  llm_model_version_override: string | null;
-  starter_messages: StarterMessage[] | null;
-  users?: string[];
-  groups: number[];
-  tool_ids: number[]; // Added tool_ids to the interface
-}
+import { Persona, PersonaCreationRequest, PersonaUpdateRequest  } from "./interfaces";
 
 function promptNameFromPersonaName(personaName: string) {
   return `default-prompt__${personaName}`;
