+++ web/src/lib/llm/utils.ts
@@ -1,4 +1,4 @@
-import { Persona } from "@/app/admin/assistants/interfaces";
+import { Persona } from "@/lib/admin/interfaces";
 import { LLMProviderDescriptor } from "@/app/admin/models/llm/interfaces";
 
 export function getFinalLLM(
